Rethrow exceptions instead of calling System.exit
With Forge, you're never supposed to call System.exit. It won't work anyway
(they disable it), and even having it in your code makes a big warning in the
log. Change this to the right way of dealing with a checked exception you
can't just propagate: wrap it in an unchecked exception.